Partial data reader, read/process file logic in main flow
Fleshed out the DataReader class. Line parsing is only described, not
implemented, but the overall logic is in place.

Defined master list of files read with their data hierarchy, id and
time loaded (so we can put it in a comment in the output file).

Added a way to read a named file, call the parser and add the result to
the master list.

Created a very simple data file for testing.