Fixed close struct and string reading off-by-one error
The close struct tag was popping the context stack twice, instead of
once and then looking at the top item.
The quoted-string reader was skipping the first character found because
it read a new character at the beginning of the loop instead of at the
end. If the first character was a quote, it would miss that and assume
the following close quote was actually an open.