Fix off-by-one error in counting term length
Changed when the count of characters used is incremented to be in the
individual places instead of attempting to figure it out in the general
case. It was causing an error when an attribute had a single-character
value followed by a symbol (eg. over18=1} )
Added ‘.’ to the list of characters excluded from attribute names
because the journal files will use it to mark a path through the
hierarchy.